+++ CoreDataServicesTests/Services/Core/EntityDescription/NSEntityDescription+EntityDescriptionTests.swift
@@ -16,11 +16,11 @@ class NSEntityDescription_EntityDescriptionTests: XCTestCase {
     override func setUp() {
         super.setUp()
         
-        ServiceManager.sharedInstance.setupModel(name: "Model", bundle: Bundle(for: ServiceManagerTests.self))
+        ServiceManager.shared.setupModel(name: "Model", bundle: Bundle(for: ServiceManagerTests.self))
     }
     
     override func tearDown() {
-        ServiceManager.sharedInstance.clear()
+        ServiceManager.shared.clear()
         
         super.tearDown()
     }
@@ -28,7 +28,7 @@ class NSEntityDescription_EntityDescriptionTests: XCTestCase {
     // MARK: - Retrieval
     
     func test_entityDescriptionFor_entityDescriptionReturned() {
-        let entityDescription = NSEntityDescription.entityDescriptionFor(entityClass: Test.self, managedObjectContext: ServiceManager.sharedInstance.mainManagedObjectContext)
+        let entityDescription = NSEntityDescription.entityDescriptionFor(entityClass: Test.self, managedObjectContext: ServiceManager.shared.mainManagedObjectContext)
         
         XCTAssertEqual(entityDescription.name!, String(describing: Test.self), "Should have an NSEntityDescription instance for \(String(describing: Test.self))")
     }
@@ -36,13 +36,13 @@ class NSEntityDescription_EntityDescriptionTests: XCTestCase {
     // MARK: - Insert
     
     func test_insertNewObjectForEntity_inserted() {
-        NSEntityDescription.insertNewObject(entityClass:Test.self, managedObjectContext: ServiceManager.sharedInstance.mainManagedObjectContext)
+        NSEntityDescription.insertNewObject(entityClass:Test.self, managedObjectContext: ServiceManager.shared.mainManagedObjectContext)
         
-        XCTAssertEqual(ServiceManager.sharedInstance.mainManagedObjectContext.retrieveEntriesCount(entityClass: Test.self), 1, "Should have one row in the \(String(describing: Test.self)) entity")
+        XCTAssertEqual(ServiceManager.shared.mainManagedObjectContext.retrieveEntriesCount(entityClass: Test.self), 1, "Should have one row in the \(String(describing: Test.self)) entity")
     }
     
     func test_insertNewObjectForEntity_returnType() {
-        let managedObject = NSEntityDescription.insertNewObject(entityClass:Test.self, managedObjectContext: ServiceManager.sharedInstance.mainManagedObjectContext)
+        let managedObject = NSEntityDescription.insertNewObject(entityClass:Test.self, managedObjectContext: ServiceManager.shared.mainManagedObjectContext)
         
         XCTAssertTrue(managedObject.isKind(of: Test.self), "Should have returned object of type: \(String(describing: Test.self))")
     }